fix : useChickenGame hook create
